Move useAuth to hooks folder
